+++ R/learner_prioritylasso_classif_priority_lasso.R
@@ -38,8 +38,8 @@ LearnerClassifPriorityLasso = R6Class("LearnerClassifPriorityLasso",
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
       param_set = ps(
-        blocks               = p_uty(default = NULL, tags = c("train", "required")),
-        type.measure         = p_fct(default = "auc", levels = c("class", "auc"), tags = c("train", "required")),
+        blocks               = p_uty(tags = c("train", "required")),
+        type.measure         = p_fct(levels = c("class", "auc"), tags = c("train", "required")),
         max.coef             = p_uty(default = NULL, tags = "train"),
         block1.penalization  = p_lgl(default = TRUE, tags = "train"),
         lambda.type          = p_fct(default = "lambda.min", levels = c("lambda.min", "lambda.1se"), tags = c("train", "predict")), #nolint