+++ NEWS.md
@@ -1,6 +1,14 @@
-# mlr3extralearners 0.7.1-9000
+# dev
 
-* Fix: Replace hardcoded `VectorDistribution`s from partykit survival learners with survival matrices (`Matdist`) (thanks to @bblodfon)
+* Added response (i.e., survival time) prediction to `aorsf` learner
+* Updated support for flexsurv v2.3
+
+# mlr3extralearners 0.8.0
+
+* Added `surv.xgboost.cox` and `surv.xgboost.aft` separate survival learners. `distr` prediction on the cox xgboost learner is now estimated via Breslow by default and aft xgboost has now in addition a `response` prediction (survival time)
+* Ported `surv.parametric` code to `survivalmodels`, changed `type` parameter to `form` to avoid conflict with survivalmodels's default parameter list
+* Fix: Replace hardcoded `VectorDistribution`s from partykit and flexsurv survival learners with survival matrices (`Matdist`) (thanks to @bblodfon)
+* Feat: Add `discrete` parameter in `surv.parametric` learner to return `Matdist` survival predictions
 * Added method `selected_features()` to CoxBoost survival learners (thanks to @bblodfon)
 * Added the Random Planted Forest Learner (thanks to @jemus42)
 * re-added the catboost learner as it was requested (was previously removed
@@ -11,6 +19,11 @@
 * Various minor doc improvements
 * Added the `distr` predict type to the `surv.cv_glmnet` and `surv.glmnet`
 learners (thanks to @bblodfon)
+* Feat: Added many new WEKA learners (thanks to @damirpolat)
+* Fix: `I` and `F` params from IBk learner are too interdependent (`I` can only be `TRUE` when `F` is `FALSE` and vice versa).
+  Combined them into one factor param `weight` that has two levels -- `I` and `F`.
+* Fix: `U` must be `FALSE` for `S` to be tunable in J48 learner.
+* Compatibility with upcoming 'paradox' release.
 
 # mlr3extralearners 0.7.1